requirements.txt
python-telegram-bot==12.8
opencv-python-headless
numpy
joblib
mercadopago
python-dotenv
pandas
urllib3==1.26.5
six==1.16.0
scikit-learn
google-generativeai